Move util.mkmetadatadir() to bodhi.tests.server.base.
I noticed that mkmetadatadir() was only used by the tests, and that
it was included in bodhi.server.util(). This commit moves the
function into the tests.
