+++ bodhi/server/util.py
@@ -133,55 +133,6 @@ def insert_in_repo(comp_type, repodata, filetype, extension, source):
     os.unlink(target_fname)
 
 
-def mkmetadatadir(path, updateinfo=None, comps=None):
-    """
-    Generate package metadata for a given directory.
-
-    If the metadata doesn't exist, then create it.
-
-    Args:
-        path (basestring): The directory to generate metadata for.
-        updateinfo (basestring or None or bool): The updateinfo to insert instead of example.
-            No updateinfo is inserted if False is passed. Passing True provides undefined
-            behavior.
-        comps (basestring or None): The comps to insert instead of example.
-    """
-    compsfile = '''<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE comps PUBLIC "-//Red Hat, Inc.//DTD Comps info//EN" "comps.dtd">
-<comps>
-  <group>
-    <id>testable</id>
-    <_name>Testable</_name>
-    <_description>comps group for testing</_description>
-    <packagelist>
-      <packagereq>testpkg</packagereq>
-    </packagelist>
-  </group>
-</comps>'''
-    updateinfofile = ''
-    if not os.path.isdir(path):
-        os.makedirs(path)
-    if not comps:
-        comps = os.path.join(path, 'comps.xml')
-        with open(comps, 'w') as f:
-            f.write(compsfile)
-    if updateinfo is None:
-        updateinfo = os.path.join(path, 'updateinfo.xml')
-        with open(updateinfo, 'w') as f:
-            f.write(updateinfofile)
-
-    subprocess.check_call(['createrepo_c',
-                           '--groupfile', 'comps.xml',
-                           '--deltas',
-                           '--xz',
-                           '--database',
-                           '--quiet',
-                           path])
-    if updateinfo is not False:
-        insert_in_repo(cr.XZ, os.path.join(path, 'repodata'), 'updateinfo', 'xml',
-                       os.path.join(path, 'updateinfo.xml'))
-
-
 def flash_log(msg):
     """
     Log the given message at debug level.
