+++ bodhi/tests/server/test_metadata.py
@@ -33,7 +33,6 @@ from bodhi.server.buildsys import (setup_buildsystem, teardown_buildsystem,
 from bodhi.server.config import config
 from bodhi.server.models import Release, Update, UpdateRequest, UpdateStatus
 from bodhi.server.metadata import UpdateInfoMetadata
-from bodhi.server.util import mkmetadatadir
 from bodhi.tests.server import base, create_update
 
 
@@ -47,7 +46,7 @@ class UpdateInfoMetadataTestCase(base.BaseTestCase):
         self.tempdir = tempfile.mkdtemp('bodhi')
         self.tempcompdir = join(self.tempdir, 'f17-updates-testing')
         self.temprepo = join(self.tempcompdir, 'compose', 'Everything', 'i386', 'os')
-        mkmetadatadir(join(self.temprepo, 'f17-updates-testing', 'i386'), updateinfo=False)
+        base.mkmetadatadir(join(self.temprepo, 'f17-updates-testing', 'i386'), updateinfo=False)
         config['cache_dir'] = os.path.join(self.tempdir, 'cache')
         os.makedirs(config['cache_dir'])
 
@@ -241,9 +240,9 @@ class TestUpdateInfoMetadata(UpdateInfoMetadataTestCase):
         os.makedirs(os.path.join(config['mash_dir'], 'f17-updates-testing'))
 
         # Initialize our temporary repo
-        mkmetadatadir(self.temprepo, updateinfo=False)
-        mkmetadatadir(join(self.tempcompdir, 'compose', 'Everything', 'source', 'tree'),
-                      updateinfo=False)
+        base.mkmetadatadir(self.temprepo, updateinfo=False)
+        base.mkmetadatadir(join(self.tempcompdir, 'compose', 'Everything', 'source', 'tree'),
+                           updateinfo=False)
         self.repodata = join(self.temprepo, 'repodata')
         assert exists(join(self.repodata, 'repomd.xml'))
 
@@ -313,9 +312,9 @@ class TestUpdateInfoMetadata(UpdateInfoMetadataTestCase):
         """
         self._test_extended_metadata(True)
         shutil.rmtree(self.temprepo)
-        mkmetadatadir(self.temprepo, updateinfo=False)
-        mkmetadatadir(join(self.tempcompdir, 'compose', 'Everything', 'source', 'tree'),
-                      updateinfo=False)
+        base.mkmetadatadir(self.temprepo, updateinfo=False)
+        base.mkmetadatadir(join(self.tempcompdir, 'compose', 'Everything', 'source', 'tree'),
+                           updateinfo=False)
         DevBuildsys.__rpms__ = []
         self._test_extended_metadata(True)
